Create main.c
This program works by reading in a user inputed Zip code as a character array. Then it converts the character array to an integer. Afterwards it finds the last digit in the zip code, saves it to a value and then chops off the last digit from the zip code. This process is repeated until the entire zipcode is converted to single digits. AFterwards the digits are all summed up to determine the checksum and that value is recorded as well. Then all of those values are fed to createBarcode where the numeric values are converted to barcode.